refactor(export): use dynamic import for subcommand
