+++ packages/opencode/src/cli/cmd/export/command.ts
@@ -1,0 +1,23 @@
+import type { Argv } from "yargs"
+import { cmd } from "../cmd"
+
+export const ExportCommand = cmd({
+  command: "export [sessionID]",
+  describe: "export session data as JSON",
+  builder: (yargs: Argv) => {
+    return yargs
+      .positional("sessionID", {
+        describe: "session id to export",
+        type: "string",
+      })
+      .option("sanitize", {
+        describe: "redact sensitive transcript and file data",
+        type: "boolean",
+      })
+  },
+  handler: async (args) => {
+    await import("./handler").then(({ handler }) => handler(args))
+  },
+})
+
+export type ExportArgs = Parameters<NonNullable<typeof ExportCommand.handler>>[0]
