+++ packages/opencode/src/cli/cmd/export/handler.ts
@@ -1,13 +1,12 @@
-import type { Argv } from "yargs"
 import { Session } from "@/session/session"
-import { MessageV2 } from "../../session/message-v2"
-import { SessionID } from "../../session/schema"
-import { cmd } from "./cmd"
-import { bootstrap } from "../bootstrap"
-import { UI } from "../ui"
+import { MessageV2 } from "../../../session/message-v2"
+import { SessionID } from "../../../session/schema"
+import { bootstrap } from "../../bootstrap"
+import { UI } from "../../ui"
 import * as prompts from "@clack/prompts"
 import { EOL } from "os"
 import { AppRuntime } from "@/effect/app-runtime"
+import type { ExportArgs } from "./command"
 
 function redact(kind: string, id: string, value: string) {
   return value.trim() ? `[redacted:${kind}:${id}]` : value
@@ -220,87 +219,72 @@ function sanitize(data: { info: Session.Info; messages: MessageV2.WithParts[] })
   }
 }
 
-export const ExportCommand = cmd({
-  command: "export [sessionID]",
-  describe: "export session data as JSON",
-  builder: (yargs: Argv) => {
-    return yargs
-      .positional("sessionID", {
-        describe: "session id to export",
-        type: "string",
+export async function handler(args: ExportArgs) {
+  await bootstrap(process.cwd(), async () => {
+    let sessionID = args.sessionID ? SessionID.make(args.sessionID) : undefined
+    process.stderr.write(`Exporting session: ${sessionID ?? "latest"}\n`)
+
+    if (!sessionID) {
+      UI.empty()
+      prompts.intro("Export session", {
+        output: process.stderr,
       })
-      .option("sanitize", {
-        describe: "redact sensitive transcript and file data",
-        type: "boolean",
-      })
-  },
-  handler: async (args) => {
-    await bootstrap(process.cwd(), async () => {
-      let sessionID = args.sessionID ? SessionID.make(args.sessionID) : undefined
-      process.stderr.write(`Exporting session: ${sessionID ?? "latest"}\n`)
 
-      if (!sessionID) {
-        UI.empty()
-        prompts.intro("Export session", {
-          output: process.stderr,
-        })
-
-        const sessions = []
-        for await (const session of Session.list()) {
-          sessions.push(session)
-        }
-
-        if (sessions.length === 0) {
-          prompts.log.error("No sessions found", {
-            output: process.stderr,
-          })
-          prompts.outro("Done", {
-            output: process.stderr,
-          })
-          return
-        }
-
-        sessions.sort((a, b) => b.time.updated - a.time.updated)
-
-        const selectedSession = await prompts.autocomplete({
-          message: "Select session to export",
-          maxItems: 10,
-          options: sessions.map((session) => ({
-            label: session.title,
-            value: session.id,
-            hint: `${new Date(session.time.updated).toLocaleString()} • ${session.id.slice(-8)}`,
-          })),
-          output: process.stderr,
-        })
-
-        if (prompts.isCancel(selectedSession)) {
-          throw new UI.CancelledError()
-        }
-
-        sessionID = selectedSession
-
-        prompts.outro("Exporting session...", {
-          output: process.stderr,
-        })
+      const sessions = []
+      for await (const session of Session.list()) {
+        sessions.push(session)
       }
 
-      try {
-        const sessionInfo = await AppRuntime.runPromise(Session.Service.use((svc) => svc.get(sessionID!)))
-        const messages = await AppRuntime.runPromise(
-          Session.Service.use((svc) => svc.messages({ sessionID: sessionInfo.id })),
-        )
-
-        const exportData = {
-          info: sessionInfo,
-          messages,
-        }
-
-        process.stdout.write(JSON.stringify(args.sanitize ? sanitize(exportData) : exportData, null, 2))
-        process.stdout.write(EOL)
-      } catch {
-        UI.error(`Session not found: ${sessionID!}`)
-        process.exit(1)
+      if (sessions.length === 0) {
+        prompts.log.error("No sessions found", {
+          output: process.stderr,
+        })
+        prompts.outro("Done", {
+          output: process.stderr,
+        })
+        return
       }
-    })
-  },
-})
+
+      sessions.sort((a, b) => b.time.updated - a.time.updated)
+
+      const selectedSession = await prompts.autocomplete({
+        message: "Select session to export",
+        maxItems: 10,
+        options: sessions.map((session) => ({
+          label: session.title,
+          value: session.id,
+          hint: `${new Date(session.time.updated).toLocaleString()} • ${session.id.slice(-8)}`,
+        })),
+        output: process.stderr,
+      })
+
+      if (prompts.isCancel(selectedSession)) {
+        throw new UI.CancelledError()
+      }
+
+      sessionID = selectedSession
+
+      prompts.outro("Exporting session...", {
+        output: process.stderr,
+      })
+    }
+
+    try {
+      const sessionInfo = await AppRuntime.runPromise(Session.Service.use((svc) => svc.get(sessionID!)))
+      const messages = await AppRuntime.runPromise(
+        Session.Service.use((svc) => svc.messages({ sessionID: sessionInfo.id })),
+      )
+
+      const exportData = {
+        info: sessionInfo,
+        messages,
+      }
+
+      process.stdout.write(JSON.stringify(args.sanitize ? sanitize(exportData) : exportData, null, 2))
+      process.stdout.write(EOL)
+    } catch {
+      UI.error(`Session not found: ${sessionID!}`)
+      process.exit(1)
+    }
+  })
+}
