+++ packages/opencode/src/index.ts
@@ -20,7 +20,7 @@ import { DebugCommand } from "./cli/cmd/debug/command"
 import { StatsCommand } from "./cli/cmd/stats"
 import { McpCommand } from "./cli/cmd/mcp"
 import { GithubCommand } from "./cli/cmd/github"
-import { ExportCommand } from "./cli/cmd/export"
+import { ExportCommand } from "./cli/cmd/export/command"
 import { ImportCommand } from "./cli/cmd/import"
 import { AttachCommand } from "./cli/cmd/tui/attach"
 import { TuiThreadCommand } from "./cli/cmd/tui/thread"
